feature: add JS Rocket sound syncronization system
