+++ src/assets/lib/jsRocket/.htaccess
@@ -1,0 +1,1 @@
+Header set Access-Control-Allow-Origin *
