+++ src/sound/JSRocket.d.ts
@@ -1,0 +1,22 @@
+// Type definitions for jsRocket 1.0.1
+// Project: https://github.com/kusma/rocket
+// http://plnkr.co/edit/gPb2rYwmydisTmpg5p7g?p=preview&preview
+/**
+ * JSRocket
+ *
+ * A sync-tracker, a tool for synchronizing music and visuals in demoscene productions.
+ *
+ * MIT license.
+ */
+declare namespace JSRocket {
+
+	function init(): void;
+
+	function SyncData(): void;
+
+	function SyncDevice(): void;
+
+	function Track(): void;
+
+	function SyncDevicePlayer(): void;
+}
