+++ src/sound/XMPlayer.d.ts
